requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
confusable-homoglyphs==3.2.0
Django==4.0.4
django-rest-registration==0.7.2
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
